# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "yes",
        "title": "twitchclone api",
        "termsOfService": "no",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
...
        "/follow/{username}": {
            "get": {
                "description": "Returns json response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "is user a follower of other user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "follower",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "followed",
                        "name": "followed",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/follow.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            },
            "post": {
                "description": "Allows the authenticated user to follow another user by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Follow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the user to follow",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username of the user to be followed",
                        "name": "following",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request: user mismatch",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: failed to follow user",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Allows the authenticated user to unfollow another user by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Unfollow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the user to unfollow",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username of the user to be unfollowed",
                        "name": "unfollowing",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unfollow successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request: user mismatch",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: failed to unfollow user",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            }
        },
        "/follows": {
            "get": {
                "description": "Lists the followees of a given follower username. If `extended=true` is passed as a query parameter, returns detailed info per followee.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Retrieve follower's follow list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Follower username whose follow list is requested",
                        "name": "follower",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If set to 'true', returns extended follow list details",
                        "name": "extended",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extended follow list response",
                        "schema": {
                            "$ref": "#/definitions/follow.ListExtendedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request: follower username is missing",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal server error while retrieving follow list",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "er.RequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "follow.GetResponse": {
            "type": "object",
            "properties": {
                "is_follower": {
                    "type": "boolean"
                }
            }
        },
        "follow.ListExtendedResponse": {
            "type": "object",
            "properties": {
                "follow_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/follow.ListExtendedResponseItem"
                    }
                }
            }
        },
        "follow.ListExtendedResponseItem": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "is_live": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "viewers": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}